Make `write` implementaion compatible with std::io::Write
If the `write`-method reuturns zero bytes written it is considered an
error in the std API. Since it is possible, that `liblzma` needs to
write out some internal buffers first before consuming more input,
`write` could return 0 in some cases. Since this seems to be normal
behaviour of `liblzma`, this patch makes sure, that all internal buffers
are written out and at least some of the input is consumed.